MAT-5211 add new APIs
